move commands functionality to individual structs and files
