+++ src/commands.rs
@@ -1,0 +1,8 @@
+// This is free and unencumbered software released into the public domain.
+
+mod external;
+mod help;
+mod help_cmd;
+pub use external::*;
+pub use help::*;
+pub use help_cmd::*;
