+++ src/commands/help.rs
@@ -1,0 +1,12 @@
+// This is free and unencumbered software released into the public domain.
+
+use crate::Result;
+
+/// Prints extensive help message, executing `help` command for each subcommand.
+pub struct HelpCommand;
+
+impl HelpCommand {
+    pub fn execute(&self) -> Result {
+        unimplemented!()
+    }
+}
